chore: improve navigation props
