+++ src/services/navigator/index.ts
@@ -1,6 +1,5 @@
 export { Navigator } from './navigator';
 
-export type { NavigationComponents, NavigatorProps } from './types';
 export { HvRouteError, HvNavigatorError, HvRenderError } from './errors';
 export {
   addStackRoute,
@@ -17,10 +16,4 @@ export {
   setSelected,
   updateRouteUrlFromState,
 } from './helpers';
-export {
-  ANCHOR_ID_SEPARATOR,
-  ID_CARD,
-  ID_MODAL,
-  KEY_MODAL,
-  NAVIGATOR_TYPE,
-} from './types';
+export { ANCHOR_ID_SEPARATOR, ID_CARD, ID_MODAL, KEY_MODAL } from './types';
