+++ src/services/navigator/types.ts
@@ -4,7 +4,6 @@ import type {
   RouteProps,
 } from 'hyperview/src/types';
 import { NavigationContainerRefContext } from '@react-navigation/native';
-import type { BottomTabBarProps as RNBottomTabBarProps } from '@react-navigation/bottom-tabs';
 
 export const ANCHOR_ID_SEPARATOR = '#';
 export const ID_CARD = 'card';
@@ -15,14 +14,6 @@ export const KEY_MODAL = 'modal';
 export const KEY_ID = 'id';
 export const KEY_TYPE = 'type';
 export const KEY_HREF = 'href';
-
-/**
- * Definition of the available navigator types
- */
-export const NAVIGATOR_TYPE = {
-  STACK: 'stack',
-  TAB: 'tab',
-};
 
 export type Props = {
   navigation?: NavigationProps;
@@ -37,24 +28,9 @@ export type Props = {
 /**
  * Mapping of screens and params for navigation
  */
-export type NavigationNavigateParams = {
+export type NavigateParams = {
   screen?: string;
-  params?: NavigationNavigateParams | RouteParams;
-};
-
-type BottomTabBarProps = RNBottomTabBarProps & {
-  id: string;
-};
-
-type BottomTabBarComponent = (props: BottomTabBarProps) => JSX.Element | null;
-
-export type NavigationComponents = {
-  BottomTabBar?: BottomTabBarComponent;
-};
-
-/* List of props available to navigators */
-export type NavigatorProps = NavigationComponents & {
-  doc: Document | undefined;
+  params?: NavigateParams | RouteParams;
 };
 
 /**
